Add consumables and Goods module, touch up $buy and $discard
