+++ Modules/Economy.cs
@@ -2,7 +2,6 @@
 [Summary("This is the hub for checking and managing your economy stuff. Wanna know how much cash you have? Or what items you have? Or do you want to check out le shoppe? It's all here.")]
 public class Economy : ModuleBase<SocketCommandContext>
 {
-    public InteractiveService Interactive { get; set; }
     public static readonly string[] CMDS_WITH_COOLDOWN = { "Deal", "Loot", "Rape", "Rob", "Scavenge",
         "Slavery", "Whore", "Bully",  "Chop", "Dig", "Farm", "Fish", "Hunt", "Mine", "Support", "Hack",
         "Daily" };
@@ -27,20 +26,6 @@ public class Economy : ModuleBase<SocketCommandContext>
             await ReplyAsync($"**{user.Sanitize()}** has **{dbUser.Cash:C2}**.", allowedMentions: Constants.MENTIONS);
 
         return CommandResult.FromSuccess();
-    }
-
-    [Alias("purchase")]
-    [Command("buy")]
-    [Summary("Buy an item from the shop.")]
-    [Remarks("$buy [item]")]
-    public async Task<RuntimeResult> Buy([Remainder] string item)
-    {
-        if (ItemSystem.tools.Any(t => t.Name.Equals(item, StringComparison.OrdinalIgnoreCase)))
-            return await ItemSystem.BuyTool(item, Context.User, Context.Guild, Context.Channel);
-        else if (ItemSystem.perks.Any(perk => perk.Name.Equals(item, StringComparison.OrdinalIgnoreCase)))
-            return await ItemSystem.BuyPerk(item, Context.User, Context.Guild, Context.Channel);
-        else
-            return CommandResult.FromError($"**{item}** is not a valid item!");
     }
 
     [Alias("cd")]
@@ -73,106 +58,6 @@ public class Economy : ModuleBase<SocketCommandContext>
             .WithTitle(user == null ? "Cooldowns" : $"{user.Sanitize()}'s Cooldowns")
             .WithDescription(description.Length > 0 ? description.ToString() : "None");
         await ReplyAsync(embed: embed.Build());
-    }
-
-    [Command("daily")]
-    [Summary("Get a daily reward.")]
-    [Remarks("$daily")]
-    [RequireCooldown("DailyCooldown", "Slow down there, turbo! It hasn't been a day yet. You've still got {0} left.")]
-    [RequireRankLevel("3")]
-    public async Task<RuntimeResult> Daily()
-    {
-        DbUser user = await DbUser.GetById(Context.Guild.Id, Context.User.Id);
-        if (user.UsingSlots)
-            return CommandResult.FromError("You appear to be currently gambling. I cannot do any transactions at the moment.");
-
-        double moneyEarned = RandomUtil.NextDouble(Constants.DAILY_MIN, Constants.DAILY_MAX);
-        double totalCash = user.Cash + moneyEarned;
-        string message = RandomUtil.Next(5) switch
-        {
-            0 => $"Your job at Pierce & Pierce is paying exceptionally well, and business is looking fantastic. **{moneyEarned:C2}** for a pretty mild day of work. That's what I'm talkin' bout.\nBalance: {totalCash:C2}",
-            1 => $"Quite a long day of disabling evil right-wingers' Discord accounts, but hey, you got yourself **{moneyEarned:C2}**. Least it's paying better than the furry shoots you were doing for quite a while.\nBalance: {totalCash:C2}",
-            2 => $"The OnlyFans money is pouring in! **{moneyEarned:C2}** from some lonely suckers just today! Thank God for giving you such a big ass.\nBalance: {totalCash:C2}",
-            3 => $"Another day of slouching on the couch and leeching off taxpayer money has gotten you **{moneyEarned:C2}**.\nBalance: {totalCash:C2}",
-            4 => $"Hot dayum! **{moneyEarned:C2}** from simp donations on your hot tub stream tonight! Your (also simp) boyfriend is gonna be ecstatic.\nBalance: {totalCash:C2}",
-            _ => ""
-        };
-
-        await user.SetCash(Context.User, totalCash);
-        user.DailyCooldown = DateTimeOffset.UtcNow.ToUnixTimeSeconds(Constants.DAILY_COOLDOWN);
-        await Context.User.NotifyAsync(Context.Channel, message);
-        return CommandResult.FromSuccess();
-    }
-
-    [Alias("sell")]
-    [Command("discard")]
-    [Summary("Discard a tool or the Pacifist perk.")]
-    [Remarks("$discard [item]")]
-    public async Task<RuntimeResult> Discard([Remainder] string item)
-    {
-        DbUser user = await DbUser.GetById(Context.Guild.Id, Context.User.Id);
-        if (user.UsingSlots)
-            return CommandResult.FromError("You appear to be currently gambling. I cannot do any transactions at the moment.");
-
-        if (item == "Pacifist")
-        {
-            if (!user.Perks.ContainsKey("Pacifist"))
-                return CommandResult.FromError("You do not have the Pacifist perk!");
-
-            user.Perks.Remove("Pacifist");
-            user.PacifistCooldown = DateTimeOffset.UtcNow.ToUnixTimeSeconds(259200);
-            await Context.User.NotifyAsync(Context.Channel, "You discarded your Pacifist perk. If you wish to buy it again, you will have to wait 3 days.");
-        }
-        else if (user.Tools.Remove(item))
-        {
-            double price = ItemSystem.GetItem(item).Price;
-            await user.SetCash(Context.User, user.Cash + price);
-            await Context.User.NotifyAsync(Context.Channel, $"You sold your {item} to some dude for **{price:C2}**.");
-        }
-        else
-        {
-            return CommandResult.FromError($"You do not have a(n) {item}!" +
-                "\n*Tip: This command is case sensitive and does not accept perks other than Pacifist.*");
-        }
-
-        return CommandResult.FromSuccess();
-    }
-
-    [Command("item")]
-    [Summary("View information on an item.")]
-    [Remarks("$item [item]")]
-    public async Task<RuntimeResult> ItemInfo([Remainder] string itemName)
-    {
-        Item item = ItemSystem.GetItem(itemName);
-        if (item is Tool tool)
-        {
-            EmbedBuilder embed = new EmbedBuilder()
-                .WithColor(Color.Red)
-                .WithTitle(tool.Name)
-                .AddField("Type", "Tool")
-                .AddField("Price", tool.Price.ToString("C2"))
-                .AddField("Cash Range", tool.Name.EndsWith("Pickaxe")
-                    ? $"{128 * tool.Mult:C2} - {256 * tool.Mult:C2}"
-                    : $"{tool.GenericMin:C2} - {tool.GenericMax:C2}");
-            await ReplyAsync(embed: embed.Build());
-        }
-        else if (item is Perk perk)
-        {
-            EmbedBuilder embed = new EmbedBuilder()
-                .WithColor(Color.Red)
-                .WithTitle(perk.Name)
-                .WithDescription(perk.Description)
-                .AddField("Type", "Perk")
-                .AddField("Price", perk.Price.ToString("C2"))
-                .AddField("Duration", TimeSpan.FromSeconds(perk.Duration).FormatCompound());
-            await ReplyAsync(embed: embed.Build());
-        }
-        else
-        {
-            return CommandResult.FromError($"**{itemName}** is not a valid item!");
-        }
-
-        return CommandResult.FromSuccess();
     }
 
     [Alias("lb")]
@@ -232,31 +117,6 @@ public class Economy : ModuleBase<SocketCommandContext>
         return CommandResult.FromSuccess();
     }
 
-    [Command("perks")]
-    [Summary("View info about your currently active perks.")]
-    [Remarks("$perks")]
-    [RequirePerk]
-    public async Task Perks()
-    {
-        DbUser user = await DbUser.GetById(Context.Guild.Id, Context.User.Id);
-        StringBuilder perksBuilder = new();
-        foreach (KeyValuePair<string, long> kvp in user.Perks.OrderBy(p => p.Key))
-        {
-            if (kvp.Value <= DateTimeOffset.UtcNow.ToUnixTimeSeconds() && kvp.Key != "Pacifist")
-                return;
-
-            Perk perk = ItemSystem.GetItem(kvp.Key) as Perk;
-            perksBuilder.AppendLine($"**{perk.Name}**: {perk.Description}" +
-                $"\nTime Left: {(perk.Name != "Pacifist" ? TimeSpan.FromSeconds(kvp.Value - DateTimeOffset.UtcNow.ToUnixTimeSeconds()).FormatCompound() : "Indefinite")}");
-        }
-
-        EmbedBuilder embed = new EmbedBuilder()
-            .WithColor(Color.Red)
-            .WithTitle("Perks")
-            .WithDescription(perksBuilder.Length > 0 ? perksBuilder.ToString() : "None");
-        await ReplyAsync(embed: embed.Build());
-    }
-
     [Alias("roles")]
     [Command("ranks")]
     [Summary("View all the ranks and their costs.")]
@@ -306,33 +166,6 @@ public class Economy : ModuleBase<SocketCommandContext>
         return CommandResult.FromSuccess();
     }
 
-    [Command("shop", RunMode = RunMode.Async)]
-    [Summary("Check out what's available for purchase in the shop.")]
-    [Remarks("$shop")]
-    public async Task Shop()
-    {
-        StringBuilder tools = new();
-        StringBuilder perks = new();
-
-        foreach (Tool tool in ItemSystem.tools)
-            tools.AppendLine($"**{tool}**: {tool.Price:C2}");
-        foreach (Perk perk in ItemSystem.perks)
-            perks.AppendLine($"**{perk.Name}**: {perk.Description}\nDuration: {TimeSpan.FromSeconds(perk.Duration).FormatCompound()}\nPrice: {perk.Price:C2}");
-
-        PageBuilder[] pages = new[]
-        {
-            new PageBuilder().WithColor(Color.Red).WithTitle("Tools").WithDescription(tools.ToString()),
-            new PageBuilder().WithColor(Color.Red).WithTitle("Perks").WithDescription(perks.ToString())
-        };
-
-        StaticPaginator paginator = new StaticPaginatorBuilder()
-            .AddUser(Context.User)
-            .WithPages(pages)
-            .Build();
-
-        await Interactive.SendPaginatorAsync(paginator, Context.Channel, resetTimeoutOnInput: true);
-    }
-
     [Alias("kms", "selfend")]
     [Command("suicide")]
     [Summary("Kill yourself.")]
@@ -377,20 +210,6 @@ public class Economy : ModuleBase<SocketCommandContext>
         }
 
         return CommandResult.FromSuccess();
-    }
-
-    [Command("tools")]
-    [Summary("Check your own or someone else's tools.")]
-    [Remarks("$tools <user>")]
-    public async Task Tools(IGuildUser user = null)
-    {
-        ulong userId = user == null ? Context.User.Id : user.Id;
-        DbUser dbUser = await DbUser.GetById(Context.Guild.Id, userId);
-        EmbedBuilder embed = new EmbedBuilder()
-            .WithColor(Color.Red)
-            .WithTitle(user == null ? "Tools" : $"{user.Sanitize()}'s Tools")
-            .WithDescription(dbUser.Tools.Count > 0 ? string.Join(", ", dbUser.Tools) : "None");
-        await ReplyAsync(embed: embed.Build());
     }
 
     private static void RestoreUserData(DbUser user, double btc, double eth, double ltc, double xrp,
